soda-data-quality/requirements.txt
# Soda Core dependencies
soda-core>=3.5.0
soda-core-postgres>=3.5.0
python-dotenv>=1.0.0

# Additional database drivers
psycopg2-binary>=2.9.7
clickhouse-connect>=0.6.0
mysql-connector-python>=8.0.0

# Utilities
pyyaml>=6.0.1
pandas>=2.0.0
setuptools>=65.0.0

# UUID generation
uuid>=1.30
